fix: restore community page rendering
